pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "gmail-inbox-summary"
dynamic = ["version"]
description = "Generate AI-powered summaries of Gmail inbox threads"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Doug Hellmann", email = "doug@doughellmann.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=0.8.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12.0",
    "python-dateutil>=2.8.0",
    "click>=8.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Documentation = "https://github.com/dhellmann/gmail-inbox-summary#readme"
Issues = "https://github.com/dhellmann/gmail-inbox-summary/issues"
Source = "https://github.com/dhellmann/gmail-inbox-summary"

[project.scripts]
gmail-summary = "gmail_summarizer.main:main"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/gmail_summarizer/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/gmail_summarizer"]

[tool.hatch.envs.default]
dependencies = [
    "coverage[toml]>=6.5",
    "pytest",
]
dev-mode = true

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
    "- coverage combine",
    "coverage report",
]
cov = [
    "test-cov",
    "cov-report",
]

[tool.hatch.envs.lint]
detached = true
dependencies = [
    "black>=23.1.0",
    "mypy>=1.0.0",
    "ruff>=0.0.243",
]

[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:src/gmail_summarizer tests}"
check = [
    "ruff check {args:.}",
    "ruff format --check {args:.}",
]
fix = [
    "ruff check --fix {args:.}",
    "ruff format {args:.}",
]
all = [
    "check",
    "typing",
]

[tool.ruff]
target-version = "py310"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "N",  # pep8-naming
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
force-single-line = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "google.*",
    "google_auth_oauthlib.*",
    "googleapiclient.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source_pkgs = ["gmail_summarizer", "tests"]
branch = true
parallel = true
omit = [
    "src/gmail_summarizer/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
    "no cov",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
